add future objects

+++ thumbnail_maker/thumbnail_maker.py
@@ -4,11 +4,12 @@ import os
 import logging
 from urllib.parse import urlparse
 from urllib.request import urlretrieve
-import threading
+from queue import Queue
+from threading import Thread
+import multiprocessing
 
 import PIL
 from PIL import Image
-
 
 FORMAT = "[%(threadName)s, %(asctime)s, %(levelname)s] %(message)s"
 logging.basicConfig(filename='logfile.log', level=logging.DEBUG, format=FORMAT)
@@ -18,52 +19,43 @@ class ThumbnailMakerService(object):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming'
         self.output_dir = self.home_dir + os.path.sep + 'outgoing'
-        self.downloaded_bytes = 0
-        self.dl_lock = threading.Lock()
-        max_concurrent_dl = 4
-        self.dl_sem = threading.Semaphore(max_concurrent_dl)
+        self.img_list = []
 
+    def download_image(self, dl_queue):
+        while not dl_queue.empty():
+            try:
+                url = dl_queue.get(block=False)
+                # download each image and save to the input dir
+                img_filename = urlparse(url).path.split('/')[-1]
+                urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+                # self.img_queue.put(img_filename)
+                self.img_list.append(img_filename)
 
-    def download_image(self, url):
-        # download each image and save to the input dir 
-        self.dl_sem.acquire()
-        try:
-            logging.info('downloading image at URL ' + url)
-            img_filename = urlparse(url).path.split('/')[-1]
-            dest_path = self.input_dir + os.path.sep + img_filename
-            urlretrieve(url, dest_path)
-            img_size = os.path.getsize(dest_path)
-            with self.dl_lock:
-                self.downloaded_bytes += img_size
-            logging.info('image [{} bytes] saved to {}'.format(img_size, img_dest))    
-        finally:
-            self.dl_sem.release()
+                dl_queue.task_done()
+            except Queue.Empty:
+                logging.info('Queue empty')
+
     def download_images(self, img_url_list):
         # validate inputs
         if not img_url_list:
             return
         os.makedirs(self.input_dir, exist_ok=True)
-        
+
         logging.info("beginning image downloads")
 
-        threads = []
         start = time.perf_counter()
         for url in img_url_list:
-           t = threading.Thread(target=self.download_image, args=(url,))
-           t.start
-           threads.append(t)
-
-        for _ in threads:
-            _.join()
-
+            # download each image and save to the input dir
+            img_filename = urlparse(url).path.split('/')[-1]
+            urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+            self.img_queue.put(img_filename)
         end = time.perf_counter()
 
+        self.img_queue.put(None)
         logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
 
     def perform_resizing(self):
         # validate inputs
-        if not os.listdir(self.input_dir):
-            return
         os.makedirs(self.output_dir, exist_ok=True)
 
         logging.info("beginning image resizing")
@@ -71,33 +63,78 @@ class ThumbnailMakerService(object):
         num_images = len(os.listdir(self.input_dir))
 
         start = time.perf_counter()
-        for filename in os.listdir(self.input_dir):
-            orig_img = Image.open(self.input_dir + os.path.sep + filename)
-            for basewidth in target_sizes:
-                img = orig_img
-                # calculate target height of the resized image to maintain the aspect ratio
-                wpercent = (basewidth / float(img.size[0]))
-                hsize = int((float(img.size[1]) * float(wpercent)))
-                # perform resizing
-                img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
-                
-                # save the resized image to the output dir with a modified file name 
-                new_filename = os.path.splitext(filename)[0] + \
-                    '_' + str(basewidth) + os.path.splitext(filename)[1]
-                img.save(self.output_dir + os.path.sep + new_filename)
+        while True:
+            filename = self.img_queue.get()
+            if filename:
+                logging.info("resizing image {}".format(filename))
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
 
-            os.remove(self.input_dir + os.path.sep + filename)
+                    # save the resized image to the output dir with a modified file name
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
+
+                os.remove(self.input_dir + os.path.sep + filename)
+                logging.info("done resizing image {}".format(filename))
+                self.img_queue.task_done()
+            else:
+                self.img_queue.task_done()
+                break
         end = time.perf_counter()
 
         logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
 
+    def resize_image(self, filename):
+        logging.info("resizing image {}".format(filename))
+        target_sizes = [32, 64, 200]
+        orig_img = Image.open(self.input_dir + os.path.sep + filename)
+        for basewidth in target_sizes:
+            img = orig_img
+            # calculate target height of the resized image to maintain the aspect ratio
+            wpercent = (basewidth / float(img.size[0]))
+            hsize = int((float(img.size[1]) * float(wpercent)))
+            # perform resizing
+            img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
+
+            # save the resized image to the output dir with a modified file name
+            new_filename = os.path.splitext(filename)[0] + \
+                '_' + str(basewidth) + os.path.splitext(filename)[1]
+            img.save(self.output_dir + os.path.sep + new_filename)
+
+        os.remove(self.input_dir + os.path.sep + filename)
+        logging.info("done resizing image {}".format(filename))
+
     def make_thumbnails(self, img_url_list):
         logging.info("START make_thumbnails")
-        start = time.perf_counter()
+        pool = multiprocessing.Pool()
 
-        self.download_images(img_url_list)
-        self.perform_resizing()
+        start = time.perf_counter()
+        dl_queue = Queue()
+
+        for img_url in img_url_list:
+            dl_queue.put(img_url)
+
+        num_dl_threads = 4
+        for _ in range(num_dl_threads):
+            t = Thread(target=self.download_image, args=(dl_queue,))
+            t.start()
+
+        dl_queue.join()
+
+        start_resize = time.perf_counter()
+        pool.map(self.resize_image, self.img_list)
+        end_resize = time.perf_counter()
 
         end = time.perf_counter()
+
+        pool.close()
+        pool.join()
+        logging.info("created {} thumbnails in {} seconds".format(len(self.img_list), end_resize - start_resize))
         logging.info("END make_thumbnails in {} seconds".format(end - start))
-
